Removing line that did not work before, because it mistakenly set the editIndex to undefined.
When these two sagas (delete row + add row) used the same function, the `index` parameter passed in was _only_ set when deleting a row (used for indicating which row to delete, but new rows are always added at the end). The thinking here was probably to set the edit index to the new row index, so I changed the code to fix that, but some tests are now failing because it is expected that the calling code handles updating the editing index (and does it conditionally).